Move PlaceholderFragment to separate file
