+++ JTime-android/src/main/res/layout/fragment_new_masjid.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.ismail_s.jtime.android.activity.NewMasjidActivity$PlaceholderFragment">
+    tools:context=".activity.PlaceholderFragment">
 
     <TextView
         android:id="@+id/section_label"
